[refactor] MainScreen 컴포넌트화 ## #️⃣연관된 이슈 X
## 📝작업 내용

> 이번 PR에서 작업한 내용을 간략히 설명해주세요(이미지 첨부 가능)

MainScreen 하나였던 파일을 컴포넌트화하였습니다. Tabbar의 디자인은 변경된 것이 아닌 이전의 것을 이용하였습니다.
