+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/main/components/PillTabBar.kt
@@ -1,0 +1,78 @@
+package com.yourssu.soomsil.usaint.screen.main.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.yourssu.soomsil.usaint.screen.main.model.TabItem
+
+@Composable
+fun PillTabBar(
+    selectedIndex: Int,
+    tabs: List<TabItem>,
+    onTabClick: (Int) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(start = 20.dp, end = 20.dp, top = 10.dp, bottom = 20.dp)
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(60.dp)
+                .background(Color.White, RoundedCornerShape(9999.dp))
+                .border(1.dp, Color(0xFFF1F5F9), RoundedCornerShape(9999.dp))
+                .padding(6.dp)
+        ) {
+            tabs.forEachIndexed { index, tab ->
+                val isSelected = index == selectedIndex
+                Column(
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxHeight()
+                        .background(
+                            if (isSelected) Color(0xFF0062FF) else Color.Transparent,
+                            RoundedCornerShape(9999.dp)
+                        )
+                        .clickable { onTabClick(index) },
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.spacedBy(4.dp, Alignment.CenterVertically)
+                ) {
+                    Icon(
+                        painter = painterResource(tab.iconRes),
+                        contentDescription = tab.label,
+                        modifier = Modifier.size(18.dp),
+                        tint = if (isSelected) Color.White else Color(0xFF9CA3AF)
+                    )
+                    Text(
+                        text = tab.label,
+                        fontSize = if (isSelected) 12.sp else 10.sp,
+                        fontWeight = if (isSelected) FontWeight.SemiBold else FontWeight.Medium,
+                        color = if (isSelected) Color.White else Color(0xFF9CA3AF)
+                    )
+                }
+            }
+        }
+    }
+}
